backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
websockets==12.0
google-generativeai==0.3.2
transformers==4.35.2
torch==2.1.1
pymilvus==2.3.4
sentence-transformers==2.2.2
python-multipart==0.0.6 